MtcpHeader: size_t->uint64_t: w/ --enable-multilib
 - Needed for './configure --enable-multilib':  Otherwise restart fails.

+++ src/mtcp/mtcp_header.h
@@ -1,6 +1,7 @@
 #ifndef MTCP_HEADER_H
 #define MTCP_HEADER_H
 
+#include <stdint.h>  // For 'uint64_t'
 #include "ldt.h"
 
 #ifdef __i386__
@@ -45,7 +46,8 @@ typedef union _MtcpHeader {
     void *saved_brk;
     void *end_of_stack;
     void *restore_addr;
-    size_t restore_size;
+    // uint64_t processinfo.h:_restoreBufLen; 64-bit aligned for 32-64-bit
+    uint64_t restore_size; /* must be 64-bit aligned */
     void *vdsoStart;
     void *vdsoEnd;
     void *vvarStart;
